[REF] core: consolidate video policy
