+++ crates/core/src/engine/room/source_policy/video/mod.rs
@@ -1,13 +1,6 @@
-mod adaptation;
-mod admission;
-mod budget;
-mod hysteresis;
 mod input;
-mod layout;
-mod planner;
 mod projection;
-mod receiver;
-mod selection;
+mod solver;
 
-pub use layout::VideoAdmissionRank;
-pub(super) use planner::append_receiver_video_policy;
+pub use solver::VideoAdmissionRank;
+pub(super) use solver::append_receiver_video_policy;
